refactor logica a main
